feat: Rename project to SparkInsight AI
Transform mcp-apache-spark-history-server into SparkInsight AI to align with brand positioning as an AI-powered evolution of the original SparkInsight project.

BREAKING CHANGES:
- Package name: mcp-apache-spark-history-server → sparkinsight-ai
- Console command: spark-mcp → sparkinsight-ai
- Module path: spark_history_mcp → sparkinsight_ai
- Import statements: from spark_history_mcp.* → from sparkinsight_ai.*

Package Structure Changes:
- Renamed: src/spark_history_mcp/ → src/sparkinsight_ai/
- Updated: pyproject.toml with new package metadata and entry points
- Updated: All Python imports across 13 source files and 7 test files

Documentation Updates:
- README.md: New title, tagline "AI-Powered Spark Analysis with MCP Integration"
- TESTING.md: Updated repository URLs and project name
- New positioning: Evolved version of SparkInsight with advanced AI capabilities

Installation Commands:
- pip install sparkinsight-ai
- uvx --from sparkinsight-ai sparkinsight-ai
- python -m sparkinsight_ai.core.main

This rebrand clearly positions the project as the AI-enhanced, MCP-enabled evolution of SparkInsight for enterprise Spark performance analysis and optimization.
