+++ pyproject.toml
@@ -1,7 +1,7 @@
 [project]
-name = "mcp-apache-spark-history-server"
+name = "sparkinsight-ai"
 dynamic = ["version"]
-description = "Model Context Protocol (MCP) server for Apache Spark History Server with job comparison and analytics"
+description = "SparkInsight AI: AI-Powered Spark Analysis with MCP Integration for intelligent performance monitoring and optimization"
 readme = "README.md"
 authors = [
     {name = "Manabu McCloskey", email = "Manabu.McCloskey@gmail.com"},
@@ -9,7 +9,7 @@ authors = [
 ]
 license = {text = "Apache-2.0"}
 requires-python = ">=3.12"
-keywords = ["spark", "mcp", "analytics", "performance", "history-server", "model-context-protocol"]
+keywords = ["spark", "ai", "mcp", "analytics", "performance", "sparkinsight", "intelligence", "model-context-protocol"]
 classifiers = [
     "Development Status :: 3 - Alpha",
     "Intended Audience :: Developers",
@@ -35,7 +35,7 @@ requires = ["hatchling", "hatch-vcs"]
 build-backend = "hatchling.build"
 
 [tool.hatch.build.targets.wheel]
-packages = ["src/spark_history_mcp"]
+packages = ["src/sparkinsight_ai"]
 
 [tool.hatch.build.targets.sdist]
 include = [
@@ -49,12 +49,12 @@ source = "vcs"
 version_scheme = "no-guess-dev"
 
 [project.scripts]
-spark-mcp = "spark_history_mcp.core.main:main"
+sparkinsight-ai = "sparkinsight_ai.core.main:main"
 
 [project.urls]
-Homepage = "https://github.com/DeepDiagnostix-AI/spark-history-server-mcp"
-Repository = "https://github.com/DeepDiagnostix-AI/spark-history-server-mcp"
-Issues = "https://github.com/DeepDiagnostix-AI/spark-history-server-mcp/issues"
+Homepage = "https://github.com/DeepDiagnostix-AI/sparkinsight-ai"
+Repository = "https://github.com/DeepDiagnostix-AI/sparkinsight-ai"
+Issues = "https://github.com/DeepDiagnostix-AI/sparkinsight-ai/issues"
 
 [tool.ruff]
 target-version = "py312"
