+++ src/sparkinsight_ai/api/spark_html_client.py
@@ -5,7 +5,7 @@ from urllib.parse import urljoin
 
 from playwright.async_api import async_playwright
 
-from spark_history_mcp.config.config import ServerConfig
+from sparkinsight_ai.config.config import ServerConfig
 
 # HTML client for Spark History Server web interface
