+++ src/sparkinsight_ai/models/mcp_types.py
