+++ src/sparkinsight_ai/models/spark_types.py
